move agents hook for creating invite email to credentials topic, hook belongs on credentials
